Refactor Package import
Previously the src prefix of the package was added, but this shouldn't be in that way, the PYTHONPATH should be set correctly indicating that src is the source root directory.
